fix: fix mistakes in code, change behavior  of the game when user authenticated or not and issues than connected with it
